order types

+++ database/migrations/2026_01_19_199999_add_attachments_to_car_request_items_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
                 ->default(false)
                 ->after('quantity');
 
-            $table->integer('attachment_type_id')
+            $table->unsignedBigInteger('attachment_type_id')
                 ->nullable()
                 ->after('is_attachments_enabled');
